major-ish refactoring; adjusted import for cli package
