+++ util/filter.go
@@ -1,6 +1,7 @@
 package util
 
 import (
+	"github.com/eveopsec/zk2s/config"
 	"github.com/vivace-io/evelib/crest"
 	"github.com/vivace-io/evelib/zkill"
 )
@@ -10,7 +11,7 @@ import (
  */
 
 // WithinFilter returns true if the kill is within the channel's filter, false otherwise
-func WithinFilter(kill *zkill.Kill, channel Channel) bool {
+func WithinFilter(kill *zkill.Kill, channel config.Channel) bool {
 	// KillID of 0 should be skipped as it means no kill was returned from RedisQ
 	if kill.KillID == 0 {
 		return false
@@ -28,7 +29,7 @@ func WithinFilter(kill *zkill.Kill, channel Channel) bool {
 }
 
 // IsLoss returns true if the kill is a loss, false otherwise
-func IsLoss(kill *zkill.Kill, channel Channel) bool {
+func IsLoss(kill *zkill.Kill, channel config.Channel) bool {
 	if characterOK(kill.Killmail.Victim.Character, channel) {
 		return true
 	}
@@ -42,7 +43,7 @@ func IsLoss(kill *zkill.Kill, channel Channel) bool {
 }
 
 // IsAwox returns true if the kill was an Awox, partial or otherwise. False if not.
-func IsAwox(kill *zkill.Kill, channel Channel) bool {
+func IsAwox(kill *zkill.Kill, channel config.Channel) bool {
 	if IsLoss(kill, channel) {
 		for a := range kill.Killmail.Attackers {
 			if characterOK(kill.Killmail.Attackers[a].Character, channel) {
@@ -59,7 +60,7 @@ func IsAwox(kill *zkill.Kill, channel Channel) bool {
 	return false
 }
 
-func valueOK(kill *zkill.Kill, channel Channel) bool {
+func valueOK(kill *zkill.Kill, channel config.Channel) bool {
 	// If kill value is within [MinimumValue, MaximumValue] return true
 	if kill.Zkb.TotalValue >= float32(channel.MinimumValue) && kill.Zkb.TotalValue <= float32(channel.MaximumValue) {
 		return true
@@ -72,7 +73,7 @@ func valueOK(kill *zkill.Kill, channel Channel) bool {
 }
 
 // returns true if ship is NOT excluded by name, false otherwise
-func shipOK(kill *zkill.Kill, channel Channel) bool {
+func shipOK(kill *zkill.Kill, channel config.Channel) bool {
 	for ship := range channel.ExcludedShips {
 		if kill.Killmail.Victim.ShipType.Name == channel.ExcludedShips[ship] {
 			return false
@@ -84,7 +85,7 @@ func shipOK(kill *zkill.Kill, channel Channel) bool {
 	return true
 }
 
-func involvedOK(kill *zkill.Kill, channel Channel) bool {
+func involvedOK(kill *zkill.Kill, channel config.Channel) bool {
 	if characterOK(kill.Killmail.Victim.Character, channel) {
 		return true
 	}
@@ -109,7 +110,7 @@ func involvedOK(kill *zkill.Kill, channel Channel) bool {
 }
 
 // returns true if passed character is in channel.IncludeCharacters, false otherwise
-func characterOK(char crest.Character, channel Channel) bool {
+func characterOK(char crest.Character, channel config.Channel) bool {
 	for c := range channel.IncludeCharacters {
 		if (char.Name == channel.IncludeCharacters[c]) || (string(char.ID) == channel.IncludeCharacters[c]) {
 			return true
@@ -119,7 +120,7 @@ func characterOK(char crest.Character, channel Channel) bool {
 }
 
 // returns true if passed corporation is in channel.IncludeCorporations, false otherwise
-func corporationOK(corp crest.Corporation, channel Channel) bool {
+func corporationOK(corp crest.Corporation, channel config.Channel) bool {
 	for c := range channel.IncludeCorporations {
 		if corp.Name == channel.IncludeCorporations[c] {
 			return true
@@ -129,7 +130,7 @@ func corporationOK(corp crest.Corporation, channel Channel) bool {
 }
 
 // returns true if passed alliance is in channel.IncludeAlliances, false otherwise
-func allianceOK(alli crest.Alliance, channel Channel) bool {
+func allianceOK(alli crest.Alliance, channel config.Channel) bool {
 	for a := range channel.IncludeAlliances {
 		if alli.Name == channel.IncludeAlliances[a] {
 			return true
